Rename json-to-record-convert sub package name from converters to jsonmapper
(cherry picked from commit c957eae0341e7b4629b8394c794b3444c25eaae7)
(cherry picked from commit c621e045729b5500c9500f4f7d3a866999cff7d7)
