+++ misc/json-to-record-converter/src/main/java/io/ballerina/jsonmapper/JsonToRecordResponse.java
@@ -16,9 +16,9 @@
  *  under the License.
  */
 
-package io.ballerina.converters;
+package io.ballerina.jsonmapper;
 
-import io.ballerina.converters.diagnostic.JsonToRecordDirectConverterDiagnostic;
+import io.ballerina.jsonmapper.diagnostic.JsonToRecordDirectConverterDiagnostic;
 
 import java.util.ArrayList;
 import java.util.List;
