+++ misc/json-to-record-converter/src/main/java/io/ballerina/jsonmapper/diagnostic/DiagnosticMessage.java
@@ -16,7 +16,7 @@
  *  under the License.
  */
 
-package io.ballerina.converters.diagnostic;
+package io.ballerina.jsonmapper.diagnostic;
 
 import io.ballerina.tools.diagnostics.DiagnosticSeverity;
