+++ misc/json-to-record-converter/src/main/java/module-info.java
@@ -1,9 +1,9 @@
-module io.ballerina.Misc.jsonToRecordConverter {
+module io.ballerina.jsonmapper {
     requires com.google.gson;
     requires io.ballerina.parser;
     requires io.ballerina.formatter.core;
     requires io.ballerina.tools.api;
     requires org.apache.commons.lang3;
 
-    exports io.ballerina.converters;
+    exports io.ballerina.jsonmapper;
 }
