+++ misc/json-to-record-converter/src/test/resources/testng.xml
@@ -22,7 +22,7 @@ under the License.
 <suite name="json-to-record-direct-converter-test-suite">
     <test name="json-to-record-direct-converter-test" parallel="false">
         <classes>
-            <class name="io.ballerina.converters.JsonToRecordDirectConverterTests"/>
+            <class name="io.ballerina.jsonmapper.JsonToRecordDirectConverterTests"/>
         </classes>
     </test>
 </suite>
