+++ settings.gradle
@@ -29,7 +29,7 @@ include(':ballerina-lang:runtime')
 include(':ballerina-lang-test')
 include(':ballerina-langlib:test')
 include(':ballerina-io-internal')
-include(':json-to-record-converter')
+include(':json-mapper')
 include(':lib-creator')
 include(':ballerina-core')
 include(':ballerina-runtime')
@@ -174,7 +174,7 @@ project(':ballerina-rt').projectDir = file('bvm/ballerina-rt')
 project(':ballerina-treegen').projectDir = file('compiler/ballerina-treegen')
 project(':ballerina-test-utils').projectDir = file('tests/ballerina-test-utils')
 project(':ballerina-io-internal').projectDir = file('misc/io-internal')
-project(':json-to-record-converter').projectDir = file('misc/json-to-record-converter')
+project(':json-mapper').projectDir = file('misc/json-to-record-converter')
 project(':lib-creator').projectDir = file('misc/lib-creator')
 project(':ballerina-core').projectDir = file('bvm/ballerina-core')
 project(':ballerina-bindgen').projectDir = file('misc/ballerina-bindgen')
